chore: rename api route (in backend and frontend)
